a little progress on it. Have the hp bars responsive and can actually play through kind of.

+++ Score/Score/CharacterDataModel.swift
@@ -13,27 +13,27 @@ class Character {
 }
 
 class Warrior: Character {
-    var hp: Int = 25
-    var maxHp: Int = 25
+    var hp: Float = 25.0
+    var maxHp: Float = 25.0
     var str: Int = 6
     var dex: Int = 2
     var agi: Int = 2
     
-    func attack() -> Int{
-        let dmg = Int(arc4random_uniform(UInt32(self.str)) + 1)
+    func attack() -> Float{
+        let dmg = Float(arc4random_uniform(UInt32(self.str)) + 1)
         return dmg
     }
 }
 
 class Thief: Character {
-    var hp: Int = 20
-    var maxHp: Int = 20
+    var hp: Float = 20.0
+    var maxHp: Float = 20.0
     var str: Int = 3
     var dex: Int = 2
     var agi: Int = 5
     
-    func attack() -> Int {
-        let dmg = Int(arc4random_uniform(UInt32(self.str)) + 1)
+    func attack() -> Float {
+        let dmg = Float(arc4random_uniform(UInt32(self.str)) + 1)
         return dmg
     }
 }
